services/email-sender/src/pyproject.toml
[project]
name = "email-sernder"
version = "0.1.0"
description = "Email sernder microservice"
requires-python = ">=3.12,<4.0"
dependencies = [
    "python-dotenv>=1.1.0,<2.0.0",
    "pydantic-settings>=2.8.1,<3.0.0",
    "backoff (==2.2.1)",
    "sentry-sdk (==2.30.0)",
    "celery>=5.5.3",
    "flower>=2.0.1",
    "sqlalchemy>=2.0.40,<3.0.0",
    "alembic>=1.15.2,<2.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "redis>=5.2.1,<6.0.0",
    "psycopg2>=2.9.10,<3.0.0",
    "httpx>=0.28.1",
    "greenlet>=3.2.0,<4.0.0",
    "aio-pika (==9.5.5)"

]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
